requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
